Refactor into package and scripts
